PHP Files added to pages
PHP files from server added. If you wish to view, you can view at 'student.csc.liv.ac.uk/~sgoperr2/#name of the page you wish to see#

PHP files wont open in a browser so if you make changes push them and ill add it to the server or save it as a html and then open on your local host.